js30: complete lesson 24
created a function fixNav that changes the navbar to fixed when the
scroll reaches the top of the navbar. when changing a element to fixed,
the navbar is considered "floating" above the page, and causes the page
to shift up. to fix this we add top padding of the ofsethight of the
navbar before it becomes fixed.